Stop generating the surrounding block on function definitions.
This change also causes the functions to be compiled when they are defined,
rather than every time they're used.

A side-effect of the change (which may or may not be good) is that functions
cannot be redefined (as they can in Python). Eval.py was updated to use unique
function names.